Backend/requirements.txt
numpy
streamlit
google-generativeai
python-dotenv
langchain
chromadb
faiss-cpu
langchain_google_genai
openai==0.28
flask